+++ src/app/components/Stack/index.tsx
@@ -1,6 +1,6 @@
 "use client";
-import { useLanguage } from "@/app/providers/LanguageContext";
-import { SiOracle, SiFirebase, SiMysql, SiFlutter, SiMongodb, SiPrisma, SiReact, SiNextdotjs, SiTailwindcss, SiDocker, SiTypescript, SiNodedotjs, SiDotnet, SiPython, SiUnity, SiPostgresql, SiStyledcomponents, SiAmazonwebservices, SiGit,  SiSass, SiExpo, SiZod, SiLinux, SiJavascript, SiExpress, SiJest,  SiVite,  } from "react-icons/si";
+import { useLanguage } from "@/providers/LanguageContext";
+import { SiOracle, SiFirebase, SiMysql, SiFlutter, SiMongodb, SiPrisma, SiReact, SiNextdotjs, SiTailwindcss, SiDocker, SiTypescript, SiNodedotjs, SiDotnet, SiPython, SiUnity, SiPostgresql, SiStyledcomponents, SiAmazonwebservices, SiGit, SiSass, SiExpo, SiZod, SiLinux, SiJavascript, SiExpress, SiJest, SiVite, } from "react-icons/si";
 import { StackCard } from "./components/StackCard";
 import { CSharpIcon } from "./components/Icons/csharpIcon";
 import { VscAzureDevops } from "react-icons/vsc";
@@ -49,7 +49,7 @@ export function Stack() {
 
                     ]} />
 
-          
+
                     {/* Já Trabalhei */}
 
                     <StackCard situation={translation[language].worked} technologies={[
@@ -80,7 +80,7 @@ export function Stack() {
                         { icon: SiLinux, name: "Linux/Wsl2" },
                         { icon: SiDocker, name: "Docker" },
                     ]} />
-                    
+
                 </div>
             </div>
         </section>
